create a new  useJoinroom hook to reuse
