+++ frontend/lib/websocket/index.ts
@@ -1,3 +1,4 @@
 export * from './websocket-manager';
 export * from './useWebSocket';
+export * from './useJoinRoom';
 export * from './constants';
