+++ frontend/modules/lucky-money/components/create-red-envelope/create-red-envelope.tsx
@@ -6,7 +6,7 @@ import { IBreadcrumb } from '@/types';
 import { ROUTES } from '@/configs/routes.config';
 import { RedEnvelopeSidebar } from './red-envelope-sidebar';
 import { CreateRedEnvelopeProvider, useCreateRedEnvelopeContext } from '../../context/CreateRedEnvelopeContext';
-import { useWebSocket } from '@/lib/websocket';
+import { useJoinRoom } from '@/lib/websocket';
 import type { WebSocketEvent } from '@/lib/websocket/websocket-manager';
 import { SOCKET_MESSAGE } from '@/lib/websocket/constants';
 import { RED_ENVELOPE_EVENT_TYPES } from '../../constants';
@@ -18,66 +18,10 @@ const breadcrumbs: IBreadcrumb[] = [
 ] as const;
 
 function CreateRedEnvelopeContent() {
-  const wsManager = useWebSocket();
   const { onRedEnvelopeStatusUpdated } = useCreateRedEnvelopeContext();
-  const joinedRef = useRef(false);
-  const joiningRef = useRef(false);
-  const intervalRef = useRef<number | null>(null);
+  const { wsManager } = useJoinRoom(SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES, true);
 
   useEffect(() => {
-    const clearJoinInterval = () => {
-      if (intervalRef.current !== null) {
-        clearInterval(intervalRef.current);
-        intervalRef.current = null;
-      }
-    };
-
-    const markJoined = () => {
-      joinedRef.current = true;
-      joiningRef.current = false;
-      clearJoinInterval();
-    };
-
-    const markLeft = () => {
-      joinedRef.current = false;
-      joiningRef.current = false;
-    };
-
-    const parseRoomEvent = (evt: unknown): { type: string; room: string } | null => {
-      if (typeof evt === 'string') {
-        const [type, r] = evt.split(':');
-        if (type && r) {
-          return { type, room: r };
-        }
-        return null;
-      }
-
-      if (evt && typeof evt === 'object') {
-        const e = evt as Record<string, unknown>;
-        if (typeof e.type === 'string' && typeof e.room === 'string') {
-          return { type: e.type, room: e.room };
-        }
-      }
-
-      return null;
-    };
-
-    const serverHandler = (evt: unknown) => {
-      const parsed = parseRoomEvent(evt);
-      if (!parsed || parsed.room !== SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES) {
-        return;
-      }
-
-      switch (parsed.type) {
-        case SOCKET_MESSAGE.SERVER_JOINED_ROOM_PREFIX:
-          markJoined();
-          break;
-        case SOCKET_MESSAGE.SERVER_LEFT_ROOM_PREFIX:
-          markLeft();
-          break;
-      }
-    };
-
     const handleRedEnvelopeListRefresh = (event: WebSocketEvent) => {
       const payload = event.payload;
       if (payload && typeof payload === 'object') {
@@ -93,49 +37,9 @@ function CreateRedEnvelopeContent() {
       }
     };
 
-    const doJoin = () => {
-      if (!wsManager) return;
-      if (joinedRef.current || joiningRef.current) return;
-      if (!wsManager.isConnected()) return;
-
-      const ok = wsManager.sendRaw(
-        JSON.stringify({
-          type: SOCKET_MESSAGE.MSG_JOIN_ROOM,
-          room: SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES,
-        })
-      );
-
-      if (ok) {
-        joiningRef.current = true;
-      }
-    };
-
-    const doLeave = () => {
-      if (!wsManager) return;
-      if (!joinedRef.current && !joiningRef.current) return;
-      wsManager.sendRaw(
-        JSON.stringify({
-          type: SOCKET_MESSAGE.MSG_LEAVE_ROOM,
-          room: SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES,
-        })
-      );
-
-      markLeft();
-    };
-
-    wsManager?.on(SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES, serverHandler);
     wsManager?.on(RED_ENVELOPE_EVENT_TYPES.RED_ENVELOPE_LIST_REFRESH, handleRedEnvelopeListRefresh);
 
-    // Join room when component mounts
-    doJoin();
-    if (!joinedRef.current && intervalRef.current === null) {
-      intervalRef.current = window.setInterval(doJoin, 500);
-    }
-
     return () => {
-      clearJoinInterval();
-      doLeave();
-      wsManager?.off(SOCKET_MESSAGE.ROOM_RED_ENVELOPE_UPDATES, serverHandler);
       wsManager?.off(RED_ENVELOPE_EVENT_TYPES.RED_ENVELOPE_LIST_REFRESH, handleRedEnvelopeListRefresh);
     };
   }, [wsManager, onRedEnvelopeStatusUpdated]);
